apps/analytics-engine/pyproject.toml
[project]
name = "analytics-engine"
version = "0.1.0"
description = "Analytics backend for portfolio management and DeFi data aggregation"
authors = [{name = "Zap Pilot Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "psycopg[binary]>=3.1.12",
    # Keep for compatibility with bare postgresql:// URLs (SQLAlchemy default driver).
    "psycopg2-binary==2.9.10",
    "httpx>=0.25.2",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.10.1",
    "numpy>=2.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = ["vulture_whitelist.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from exception chaining
    "C901",  # too complex
    "C401",  # unnecessary generator
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict = true
disallow_untyped_defs = true
files = ["src/"]
exclude = ["tests/", "test_*.py"]
explicit_package_bases = true
namespace_packages = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: PostgreSQL integration tests requiring DATABASE_INTEGRATION_URL (deselect with '-m \"not integration\"')",
    "postgres: Tests requiring PostgreSQL-specific features (::numeric casts, window functions)",
]
filterwarnings = [
    "error",
    # pytest-asyncio infrastructure: event loop and unix socket cleanup during teardown
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "test_*.py"]

[tool.coverage.report]
fail_under = 99
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "^\\s*\\.\\.\\.$",
]

[dependency-groups]
dev = [
    "asyncpg>=0.31.0",
    "greenlet>=3.2.3",
    "mypy>=1.16.1",
    "pylint>=3.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.1",
    "pre-commit>=3.5.0",
    "vulture>=2.11",
    "matplotlib>=3.8.0",
]

[tool.vulture]
min_confidence = 80
paths = ["src", "vulture_whitelist.py"]
sort_by_size = true
verbose = false

[tool.pylint.main]
# Analyze import fallback blocks
analyse-fallback-blocks = false
# Pylint only cares about Python files
ignore = ["CVS", ".git", "__pycache__", ".pytest_cache", ".mypy_cache", ".ruff_cache"]
# Files or directories matching regex patterns are skipped
ignore-patterns = ["^\\.#", "test_.*\\.py", "conftest\\.py"]
# Python code to execute to configure pylint
init-hook = ""
# Number of processes to use (0=auto-detect, 1=sequential)
jobs = 0
# Control amount of debugging information
limit-inference-results = 100
# Always return 0 exit code (useful for pre-commit hooks)
exit-zero = false
# Pickle collected data for later comparisons
persistent = true
# Minimum Python version
py-version = "3.11"

[tool.pylint.messages_control]
# Only run the similarities checker, disable all other checks
disable = "all"
enable = "duplicate-code"

[tool.pylint.similarities]
# Minimum lines number of a similarity (moderate sensitivity: 6-8 lines)
min-similarity-lines = 7
# Ignore comments when computing similarities
ignore-comments = true
# Ignore docstrings when computing similarities (changed to false for better duplicate detection)
ignore-docstrings = false
# Ignore imports when computing similarities
ignore-imports = true
# Signatures are removed from the similarity computation
ignore-signatures = true
#
# NOTE: Pylint 4.0.2 has a known bug where duplicate-code (R0801) detection
# doesn't produce output even with correct configuration. This affects both
# parallel (jobs=0) and single-threaded (jobs=1) modes. Use check_duplicates.py
# as a workaround until this is resolved. See: https://github.com/pylint-dev/pylint/issues/3314

[tool.pylint.reports]
# Set to yes to activate formatted output reports
output-format = "colorized"
# Tells whether to display full report or only messages
reports = false
# Score threshold for code quality (not relevant for duplicate-only mode)
fail-under = 10
# Activate score evaluation (not relevant for duplicate-only mode)
score = false
